Quancao

+++ FPT.TeamMatching/FPT.TeamMatching.Domain/Models/Requests/Commands/Rates/RateUpdateCommand.cs
@@ -8,7 +8,7 @@ public class RateUpdateCommand : UpdateCommand
 
     public Guid? RateById { get; set; }
 
-    public int NumbOfStar { get; set; }
+    public double NumbOfStar { get; set; }
 
     public string? Content { get; set; }
 }
